Revert "FROMLIST: binder: move binder_alloc to separate file"
This reverts commit c3857183e2afabf22a53645699f554bfb1962a83.
